% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_2d_explore.R
\name{GUI_2d_explore}
\alias{GUI_2d_explore}
\title{A GUI for exploring 2d data}
\usage{
GUI_2d_explore(data)
}
\arguments{
\item{data}{Use awesome data.table to get going...}
}
\description{
Explore data stored in data.tables!
}
